+++ react/CHANGELOG.md
@@ -4,4 +4,5 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.0.0-beta
 
-* Moved Javascript config to a monorepo
+* Moved React config to a monorepo
+* Removed eslint-config-wolox as a dependency